+++ lib/src/features/sectors/presentation/add_update_sector/add_update_sector_form.dart
@@ -1,0 +1,8 @@
+class AddUpdateSectorForm extends StatelessWidget {
+  const AddUpdateSectorForm({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
